Protocol specific persistent and entity state
