+++ plt/plt-block-state/src/entity.rs
@@ -1,8 +1,5 @@
 //! Stateful entity model for block state
 
+pub mod accounts;
 pub mod block_state;
 pub mod protocol_level_tokens;
-pub mod accounts;
-
-use concordium_base::contracts_common::AccountAddress;
-
