+++ plt/plt-block-state/src/persistent/block_state/p11.rs
@@ -3,7 +3,7 @@ use crate::block_state::blob_store::{BlobStoreLoad, BlobStoreStore, Loadable, St
 use crate::block_state::cacheable::Cacheable;
 use crate::block_state::hash::Hashable;
 use crate::block_state_interface::{BlockStateFailure, BlockStateResult};
-use crate::persistent::protocol_level_tokens::PersistentPlTokens;
+use crate::persistent::protocol_level_tokens::p9::PersistentPlTokensP9;
 use concordium_base::common::Buffer;
 use concordium_base::hashes::Hash;
 use std::io::Read;
@@ -12,7 +12,7 @@ use std::io::Read;
 #[derive(Debug, Clone)]
 pub struct PersistentBlockStateP11 {
     /// Protocol-level tokens
-    pub tokens: HashedCacheableRef<PersistentPlTokens>,
+    pub tokens: HashedCacheableRef<PersistentPlTokensP9>,
 }
 
 impl Loadable for PersistentBlockStateP11 {
