+++ plt/plt-block-state/src/persistent/block_state/p9.rs
@@ -2,7 +2,7 @@ use crate::block_state::blob_store::{BlobStoreLoad, BlobStoreStore, Loadable, St
 use crate::block_state::cacheable::Cacheable;
 use crate::block_state::hash::Hashable;
 use crate::block_state_interface::{BlockStateFailure, BlockStateResult};
-use crate::persistent::protocol_level_tokens::PersistentPlTokens;
+use crate::persistent::protocol_level_tokens::p9::PersistentPlTokensP9;
 use concordium_base::common::Buffer;
 use concordium_base::hashes::Hash;
 use std::io::Read;
@@ -11,7 +11,7 @@ use std::io::Read;
 #[derive(Debug, Clone)]
 pub struct PersistentBlockStateP9 {
     /// Protocol-level tokens
-    pub tokens: PersistentPlTokens,
+    pub tokens: PersistentPlTokensP9,
 }
 
 impl Loadable for PersistentBlockStateP9 {
